Cart functionality implemented
Created a cart reducer which handles the cart state, also added the cartItems to local storage so even when the broswer is closed the user can return to the site and the products will still be waiting for them in the cart. The cart screen cycles though each item in the cart and displays the product data using .map. You can also update the quantity of an item inside the cart screen or completely remove an item which will also update the local storage.